Add new dropdown to app bar
