+++ src/app-shell.js
@@ -1,11 +1,14 @@
 import React from 'react'
+import { Container } from 'semantic-ui-react'
 import { AppBar } from './components'
 
 function AppShell({ children }) {
     return (
         <>
             <AppBar />
-            <main>{children}</main>
+            <main>
+                <Container>{children}</Container>
+            </main>
         </>
     )
 }
