+++ src/components/app-bar/index.js
@@ -1,7 +1,9 @@
 import React from 'react'
 import { useHistory } from 'react-router-dom'
-import { Menu, Image, Dropdown, Button } from 'semantic-ui-react'
-import { useAuth } from '../context/auth-context'
+import { Menu, Image, Button } from 'semantic-ui-react'
+import AvatarDropdown from './avatar-dropdown'
+import NewDropdown from './new-dropdown'
+import { useAuth } from '../../context/auth-context'
 
 function AppBar() {
     const history = useHistory()
@@ -22,7 +24,28 @@ function AppBar() {
                     </div>
                 </Menu.Item>
                 {isAuthenticated() ? (
-                    <AvatarMenu />
+                    <>
+                        <Menu.Item onClick={() => history.push('/members')}>
+                            Members
+                        </Menu.Item>
+                        <Menu.Item onClick={() => history.push('/admin')}>
+                            Admin
+                        </Menu.Item>
+                        <Menu.Menu
+                            position="right"
+                            style={{
+                                marginBottom: 'auto',
+                                marginTop: 'auto',
+                            }}
+                        >
+                            <Menu.Item>
+                                <NewDropdown />
+                            </Menu.Item>
+                            <Menu.Item>
+                                <AvatarDropdown />
+                            </Menu.Item>
+                        </Menu.Menu>
+                    </>
                 ) : (
                     <Menu.Item position="right">
                         <Button
@@ -38,47 +61,6 @@ function AppBar() {
                 )}
             </Menu>
         </>
-    )
-}
-
-function AvatarMenu() {
-    const { logout, authState } = useAuth()
-
-    const trigger = (
-        <>
-            <Image
-                avatar
-                src={authState.userInfo.profileImageUrl}
-                alt="profile image"
-            />
-            <span>{authState.userInfo.firstName}</span>
-        </>
-    )
-
-    return (
-        <Menu.Menu
-            position="right"
-            style={{
-                paddingRight: '10px',
-                marginBottom: 'auto',
-                marginTop: 'auto',
-            }}
-        >
-            <Dropdown
-                pointing
-                trigger={trigger}
-                image={{
-                    avatar: true,
-                    src: authState.userInfo.profileImageUrl,
-                }}
-            >
-                <Dropdown.Menu>
-                    <Dropdown.Item onClick={() => logout()}>
-                        Logout
-                    </Dropdown.Item>
-                </Dropdown.Menu>
-            </Dropdown>
-        </Menu.Menu>
     )
 }
